Drop PipeEvaluator; let ClashEvaluator extend Ghci
